+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Updated to support OTP 24 and no longer support OTP < 22.1
 
+* `Assent.Strategy.OIDC` now handles missing `id_token` in token params
+
 ## v0.1.22 (2021-01-08)
 
 * `Assent.Strategy.OAuth2.fetch_user/4` now accepts headers in arguments